added comments for code review
